requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.4.5.2
chardet==3.0.4
html2text==2020.1.16
html5lib==1.0.1
idna==2.9
lxml==4.5.1
python-dateutil==2.8.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.9
webencodings==0.5.1
